Completed implementation of ReSTIR GI algorithm
Tested on the Cornell Box and Fireplace scene. In the former, the
algorithm performs well when the spatial sampling radius is roughly 100
pixels (we have yet to implement adaptive control over this). It works
fairly well in the Fireplace scene as well, but struggles with
non-diffuse surfaces like the glass and mirror.